Refactor the tv test app structure to make consistency as the testapp
Refactor the tv test app structure to make consistency as the testapp
