+++ code/assurance-tv-testapp/src/main/java/com/adobe/marketing/mobile/assurance/assurancetvtestapp/ui/views/about/views/AboutView.kt
@@ -1,0 +1,61 @@
+/*
+ * Copyright 2025 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+package com.adobe.marketing.mobile.assurance.assurancetvtestapp.ui.views.about.views
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Info
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.tv.material3.*
+
+@OptIn(ExperimentalTvMaterial3Api::class)
+@Composable
+fun AboutView() {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(16.dp),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        androidx.tv.material3.Icon(
+            imageVector = Icons.Default.Info,
+            contentDescription = "About",
+            tint = Color(0xFFFA0F00),
+            modifier = Modifier.size(64.dp)
+        )
+        Spacer(modifier = Modifier.height(16.dp))
+        androidx.tv.material3.Text(
+            text = "About",
+            style = androidx.tv.material3.MaterialTheme.typography.headlineMedium,
+            color = Color.White,
+            fontWeight = FontWeight.Bold
+        )
+        Spacer(modifier = Modifier.height(8.dp))
+        androidx.tv.material3.Text(
+            text = "Android TV Test App",
+            style = androidx.tv.material3.MaterialTheme.typography.bodyLarge,
+            color = Color.White.copy(alpha = 0.7f)
+        )
+        Spacer(modifier = Modifier.height(4.dp))
+        androidx.tv.material3.Text(
+            text = "Version 1.0.0",
+            style = androidx.tv.material3.MaterialTheme.typography.bodyMedium,
+            color = Color.White.copy(alpha = 0.5f)
+        )
+    }
+}
