+++ code/assurance-tv-testapp/src/main/java/com/adobe/marketing/mobile/assurance/assurancetvtestapp/ui/views/components/theme/AppSizes.kt
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2025 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+package com.adobe.marketing.mobile.assurance.assurancetvtestapp.ui.views.components.theme
+
+import androidx.compose.ui.unit.dp
+
+object AppSizes {
+    // Icon Sizes
+    val IconTiny = 6.dp
+    val IconSmall = 12.dp
+    val IconMedium = 20.dp
+    val IconLarge = 24.dp
+    val IconExtraLarge = 36.dp
+    val IconXXL = 48.dp
+    val IconXXXL = 64.dp
+    
+    // Button/Card Heights
+    val QuickActionCardHeight = 80.dp
+    val NavigationItemHeight = 56.dp
+    val VideoCardHeight = 180.dp
+    
+    // Loading Indicator
+    val LoadingIndicatorSize = 48.dp
+    val LoadingIndicatorStroke = 4.dp
+    
+    // Navigation
+    val NavigationAccentWidth = 4.dp
+    val NavigationAccentHeight = 32.dp
+    
+    // Borders and Lines
+    val AccentLineHeight = 3.dp
+    val AccentLineWidth = 60.dp
+}
