+++ code/assurance-tv-testapp/src/main/java/com/adobe/marketing/mobile/assurance/assurancetvtestapp/ui/views/components/theme/AppSpacing.kt
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2025 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+package com.adobe.marketing.mobile.assurance.assurancetvtestapp.ui.views.components.theme
+
+import androidx.compose.ui.unit.dp
+
+object AppSpacing {
+    // Padding Values
+    val ExtraSmall = 4.dp
+    val Small = 8.dp
+    val Medium = 12.dp
+    val Large = 16.dp
+    val ExtraLarge = 24.dp
+    val XXL = 32.dp
+    val XXXL = 48.dp
+    
+    // Common Spacer Heights
+    val SpacerTiny = 2.dp
+    val SpacerSmall = 4.dp
+    val SpacerMedium = 8.dp
+    val SpacerLarge = 16.dp
+    val SpacerExtraLarge = 24.dp
+    
+    // Layout Specific
+    val CardPadding = Large
+    val SectionSpacing = ExtraLarge
+    val ContentPadding = Large
+    val NavigationPadding = 20.dp
+    val SidebarWidth = 240.dp
+}
