+++ code/assurance-tv-testapp/src/main/java/com/adobe/marketing/mobile/assurance/assurancetvtestapp/ui/views/home/views/HomeScreen.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2025 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+package com.adobe.marketing.mobile.assurance.assurancetvtestapp.ui.views.home.views
+
+import androidx.compose.runtime.Composable
+import com.adobe.marketing.mobile.assurance.assurancetvtestapp.ui.views.videos.model.Video
+
+@Composable
+fun HomeScreen(
+    videos: List<Video>,
+    isLoading: Boolean,
+    error: String?,
+    onVideoClick: (Video) -> Unit,
+    onRetry: () -> Unit = {}
+) {
+    HomeView(
+        videos = videos,
+        isLoading = isLoading,
+        error = error,
+        onVideoClick = onVideoClick,
+        onRetry = onRetry
+    )
+}
