--- code/assurance-tv-testapp/src/main/java/com/adobe/marketing/mobile/assurancetvtestapp/ui/components/ErrorView.kt
@@ -1,89 +1,0 @@
-/*
- * Copyright 2025 Adobe. All rights reserved.
- * This file is licensed to you under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License. You may obtain a copy
- * of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
- * OF ANY KIND, either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-
-package com.adobe.marketing.mobile.assurancetvtestapp.ui.components
-
-import androidx.compose.foundation.layout.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Info
-import androidx.compose.material.icons.filled.PlayArrow
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import androidx.tv.material3.*
-
-@OptIn(ExperimentalTvMaterial3Api::class)
-@Composable
-fun ErrorView(
-    error: String,
-    onRetry: (() -> Unit)? = null,
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .padding(32.dp),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        androidx.tv.material3.Icon(
-            imageVector = Icons.Default.Info,
-            contentDescription = "Error",
-            tint = Color(0xFFFA0F00),
-            modifier = Modifier.size(64.dp)
-        )
-        
-        Spacer(modifier = Modifier.height(16.dp))
-        
-        androidx.tv.material3.Text(
-            text = "Something went wrong",
-            style = androidx.tv.material3.MaterialTheme.typography.headlineMedium,
-            color = Color.White,
-            fontWeight = FontWeight.Bold
-        )
-        
-        Spacer(modifier = Modifier.height(8.dp))
-        
-        androidx.tv.material3.Text(
-            text = error,
-            style = androidx.tv.material3.MaterialTheme.typography.bodyLarge,
-            color = Color.White.copy(alpha = 0.7f),
-            textAlign = TextAlign.Center,
-            modifier = Modifier.padding(horizontal = 16.dp)
-        )
-        
-        if (onRetry != null) {
-            Spacer(modifier = Modifier.height(24.dp))
-            
-            androidx.tv.material3.Button(
-                onClick = onRetry,
-                colors = androidx.tv.material3.ButtonDefaults.colors(
-                    containerColor = Color(0xFFFA0F00)
-                )
-            ) {
-                androidx.tv.material3.Icon(
-                    imageVector = Icons.Default.PlayArrow,
-                    contentDescription = "Retry",
-                    modifier = Modifier.size(20.dp)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                androidx.tv.material3.Text(
-                    text = "Retry",
-                    color = Color.White
-                )
-            }
-        }
-    }
-}
