--- code/assurance-tv-testapp/src/main/java/com/adobe/marketing/mobile/assurancetvtestapp/videosView/network/VideoService.kt
@@ -1,24 +1,0 @@
-/*
- * Copyright 2025 Adobe. All rights reserved.
- * This file is licensed to you under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License. You may obtain a copy
- * of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
- * OF ANY KIND, either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-
-package com.adobe.marketing.mobile.assurancetvtestapp.videosView.network
-
-import com.adobe.marketing.mobile.assurancetvtestapp.videosView.model.Video
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface VideoService {
-    @GET("videos/popular")
-    suspend fun getPopularVideos(
-        @Query("page") page: Int = 1,
-        @Query("per_page") perPage: Int = 20
-    ): List<Video>
-}
