--- code/assurance-tv-testapp/src/main/res/animator/tv_button_focus_animator.xml
@@ -1,1 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><set xmlns:android="http://schemas.android.com/apk/res/android"><objectAnimator android:duration="150" android:interpolator="@android:interpolator/fast_out_slow_in" android:propertyName="scaleX" android:valueFrom="1.0" android:valueTo="1.1" android:valueType="floatType" /><objectAnimator android:duration="150" android:interpolator="@android:interpolator/fast_out_slow_in" android:propertyName="scaleY" android:valueFrom="1.0" android:valueTo="1.1" android:valueType="floatType" /></set>
