--- code/assurance-tv-testapp/src/main/res/drawable/tv_selector_background.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
- * Copyright 2025 Adobe. All rights reserved.
- * This file is licensed to you under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License. You may obtain a copy
- * of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
- * OF ANY KIND, either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- -->
-
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <!-- Focused state -->
-    <item android:state_focused="true">
-        <shape android:shape="rectangle">
-            <solid android:color="@color/tv_card_focused" />
-            <stroke
-                android:width="@dimen/tv_focus_border_width"
-                android:color="@color/adobe_red" />
-            <corners android:radius="@dimen/tv_focus_corner_radius" />
-        </shape>
-    </item>
-    
-    <!-- Selected state -->
-    <item android:state_selected="true">
-        <shape android:shape="rectangle">
-            <solid android:color="@color/tv_card_selected" />
-            <corners android:radius="@dimen/tv_focus_corner_radius" />
-        </shape>
-    </item>
-    
-    <!-- Pressed state -->
-    <item android:state_pressed="true">
-        <shape android:shape="rectangle">
-            <solid android:color="@color/adobe_red_dark" />
-            <corners android:radius="@dimen/tv_focus_corner_radius" />
-        </shape>
-    </item>
-    
-    <!-- Default state -->
-    <item>
-        <shape android:shape="rectangle">
-            <solid android:color="@color/tv_card_background" />
-            <corners android:radius="@dimen/tv_focus_corner_radius" />
-        </shape>
-    </item>
-</selector>
